# api/openapi.yaml
openapi: 3.0.3
info:
  title: Delim Gateway API
  version: 1.0.0
paths:
...
  /api/v1/auth/max:
    post:
      summary: Create a session from verified MAX initData
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/MAXLoginRequest"}}}
      responses:
        "200": {description: Session created, content: {application/json: {schema: {$ref: "#/components/schemas/MAXLoginResponse"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "500": {$ref: "#/components/responses/InternalError"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
...
  /api/v1/max-subscription:
    get:
      summary: Get the current user's personal MAX notification subscription
      security: [{BearerAuth: []}]
      responses:
        "200": {description: Subscription state, content: {application/json: {schema: {$ref: "#/components/schemas/MAXSubscription"}}}}
    post:
      summary: Get the personal bot link required to connect notifications
      security: [{BearerAuth: []}]
      responses:
        "200": {description: Bot link, content: {application/json: {schema: {$ref: "#/components/schemas/MAXSubscription"}}}}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
    delete:
      summary: Disable the current user's personal MAX notifications
      security: [{BearerAuth: []}]
      responses:
        "204": {description: Disabled}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
...
components:
  securitySchemes:
    BearerAuth: {type: http, scheme: bearer}
...
  responses:
    BadRequest: {description: Invalid request, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
    Unauthorized: {description: Authentication failed, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
...
    ServiceUnavailable: {description: Service unavailable, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
...
    InternalError: {description: Internal server error, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
  schemas:
...
    MAXLoginRequest:
      type: object
      additionalProperties: false
      required: [init_data]
      properties: {init_data: {type: string}}
    MAXLoginResponse:
      type: object
      required: [token, expires_in, user]
      properties:
        token: {type: string}
        expires_in: {type: integer, format: int64}
        user: {$ref: "#/components/schemas/MAXLoginUser"}
        start_param: {type: string}
        invite: {$ref: "#/components/schemas/MAXLoginInvite"}
        launch: {$ref: "#/components/schemas/MAXLaunch"}
    MAXLaunch:
      type: object
      required: [action]
      properties:
        action: {type: string, enum: [group, new_expense, balance, expense, settlement]}
        group_id: {type: integer, format: int64}
        entity_id: {type: integer, format: int64}
    MAXLoginUser:
      type: object
      required: [id, max_user_id]
      properties:
        id: {type: integer, format: int64}
        max_user_id: {type: integer, format: int64}
    MAXLoginInvite:
      type: object
      required: [status]
      properties:
        status: {type: string, enum: [invalid, expired, unavailable, join_failed, joined]}
        group_id: {type: integer, format: int64}
        expires_at: {type: string, format: date-time}
...
    MAXSubscription:
      type: object
      required: [connected]
      properties:
        connected: {type: boolean}
        bot_url: {type: string, format: uri}
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}